Move the ExponentiatedGradientResult into its own file
