+++ fairlearn/reductions/_exponentiated_gradient/__init__.py
@@ -4,7 +4,7 @@
 """Package for Exponentiated Gradient."""
 
 from .exponentiated_gradient import ExponentiatedGradient  # noqa: F401
-from .exponentiated_gradient import ExponentiatedGradientResult  # noqa: F401
+from ._exponentiated_gradient_result import ExponentiatedGradientResult  # noqa: F401
 
 __all__ = [
     "ExponentiatedGradient",
